dev: move breadcrumb generation to model
